分离_asyncToGenerator到单独模块，并添加单元测试
