+++ test/async.test.js
@@ -1,7 +1,7 @@
 const fs = require("fs");
 const path=require('path');
 const assert=require('assert');
-const {Promise_fromCallback,Promise_fromStandard}=require('../src/async');
+const {_asyncToGenerator,Promise_fromCallback,Promise_fromStandard}=require('../src/async');
 
 
 describe('test async.js',function(){
@@ -63,6 +63,23 @@ describe('test async.js',function(){
                     assert.fail(`callback(err,data) should resolve when err is null/undefined .etc`);
                 }
             );
+        });
+    });
+
+    describe('test #_asyncToGenerator',function(){
+        
+        it('simple test',function(){
+            let i=0;
+            function* fn(){
+                yield ++i;
+                yield ++i;
+                yield ++i;
+                yield ++i;
+            }
+            const f=_asyncToGenerator(fn);
+            return f().then(final=>{
+                assert.equal(i,4,"i should equals 4 when final yield executed");
+            });
         });
     });
 });
